% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleHaplo.R
\name{haplotypes}
\alias{haplotypes}
\title{Create a Table With the Possible Haplotypes From a Table of Polyploid Genotypes}
\usage{
haplotypes(table, ploidy = 3)
}
\arguments{
\item{table}{Table with the polyploid loci, one organisms per row}

\item{ploidy}{Ploidy level of the organism}
}
\value{
Table of haploid data from the polyploid table
}
\description{
This function loads a table with a set of polyploid markers and separates its possible haplotypes.
}
