% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationPVR.R
\name{nPVR}
\alias{nPVR}
\title{Calculates a Phylogenetic Correlation From a Network Using PVR}
\usage{
nPVR(pop, traits, dismet = NULL, pvrmet = NULL, gendist = NULL)
}
\arguments{
\item{pop}{Population in a poppr format. Can be imported from a Genalex formatted file.}

\item{traits}{Vector of the phenotypic characteristics of the members of the populations.}

\item{dismet}{Metric for distance in the network. Defaults to normalized average commute time.}

\item{pvrmet}{Metric for the correlation coefficient. Defaults to Moran's I}

\item{gendist}{Metric for the genetic distance to elaborate the MSN. Defaults to Bruvo distance. Other options include: "nei" for Nei's distance, "bitwise" for Bitwise distance, "absolute" for absolute distance, "rogers" for Roger's distance, and "reynolds" for Reynold's distance.}
}
\value{
Returns the eigenvalue correlation between the traits given and the distance between tne nodes in the network
}
\description{
This function calculates the phylogenetic correlation between the genetic distance calculated from a minimal spanning network. This uses the average commute time to calculate the distance between the nodes of the netowrk and it takes phenotypic information from the user. Each individual must be unique (incompatible with various individuals with the same genotype).
}
