% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleHaplo.R
\name{remotion}
\alias{remotion}
\title{Removes the Members of a Given Population}
\usage{
remotion(table, value)
}
\arguments{
\item{table}{Table where the isolates/organisms are.}

\item{value}{The name or list of names of the population(s) to be removed.}
}
\description{
This function takes as parameter the name of a population and removes all the individuals of that population.
}
