% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleHaplo.R
\name{remzeroes}
\alias{remzeroes}
\title{Remove All Individuals That Have a Certain Amount of 0s as Markers}
\usage{
remzeroes(table, cutoff = 6)
}
\arguments{
\item{table}{The table containing the individuals and the genotypes.}

\item{cutoff}{The cutoff value for missing loci. Defaults to 6.}
}
\value{
Table of the genotypes without the individuals with missing loci above of the cutoff.
}
\description{
Takes as parameter a cutoff value and removes all individuals which have more missing loci than the cutoff value.
}
