% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleHaplo.R
\name{rndhaplo}
\alias{rndhaplo}
\title{Create a Table With Shuffled Haplotypes From a Table of Polyploid Genotypes}
\usage{
rndhaplo(table, ploidy = 3)
}
\arguments{
\item{table}{Table with the polyploid loci, one organisms per row}

\item{ploidy}{Ploidy level of the organism}
}
\value{
Table of haploid data from the polyploid table
}
\description{
This function loads a table with a set of polyploid markers, separates its loci by the ploidy level and then randomly shuffles the markers generating a hypothetical population of randomly generated haploid individuals.
}
