% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleHaplo.R
\name{samhaplo}
\alias{samhaplo}
\title{Randomly Shuffles the Loci From the Haplotypes Within a Population}
\usage{
samhaplo(table)
}
\arguments{
\item{table}{Table with a column of the population each sample belongs to and with the haploid loci for each marker}
}
\value{
Table of newly generated individuals from the reshuffling of the genetic markers
}
\description{
This function takes a table of haplotypes with a population column and shuffles the loci within each population.
}
