% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationPVR.R
\name{sim2gen}
\alias{sim2gen}
\title{Converts the Results of an Evolutionary Simulation into a Genalex File.}
\usage{
sim2gen(x, fname)
}
\arguments{
\item{x}{Population resulting from the evolutionary simulation.}

\item{fname}{File name for the resultingg Genalex File}
}
\description{
Converts the resulting population of an evolutionary simulation from the Simulation function of this same package.
}
