% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table2Genepop.R
\name{tab2genepop}
\alias{tab2genepop}
\title{Converts a Table of Genetic Markers Into a Genepop File}
\usage{
tab2genepop(table, filen, abc = TRUE, sexratio = 1, header = NULL)
}
\arguments{
\item{table}{Table with the genetic markers}

\item{filen}{Output file name}

\item{abc}{Boolean. If true it writes a first line with the male to female sex ratio of the population. Defaults to True}

\item{sexratio}{Male-to-Female ratio. If abc is False, it is ignored}

\item{header}{Additional information to be put in the header of the GenePop file. Defaults to NULL}
}
\value{
Writes a GenePop file on the working directory
}
\description{
This function loads a table with a set of haploid markers and generates a GenePop file with this information
}
